requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
backoff==2.2.1
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.7.2
